feat: move pagetitle
