+++ packages/web/src/features/landing/frames/MainPageMainFrame.tsx
@@ -3,9 +3,9 @@
 import React from "react";
 import styled from "styled-components";
 
-import PageTitle from "@sparcs-clubs/web/common/components/PageTitle";
 import colors from "@sparcs-clubs/web/styles/themes/colors";
 
+import Typography from "@sparcs-clubs/web/common/components/Typography";
 import NoticeSectionFrame from "./NoticeSectionFrame";
 import ServiceSectionFrame from "./ServiceSectionFrame";
 
@@ -29,13 +29,27 @@ const NoticeAndServiceWrapper = styled.div`
 const MainPageMainFrame: React.FC = () => (
   <MainPageMainFrameInner>
     <PageTitleWrapper>
-      <PageTitle>
+      <Typography
+        ff="PRETENDARD"
+        fs={32}
+        lh={48}
+        fw="SEMIBOLD"
+        color="BLACK"
+        style={{ width: "fit-content" }}
+      >
         <span style={{ color: colors.PRIMARY }}>동아리</span>의 모든 것을 한번에
-      </PageTitle>
-      <PageTitle>
+      </Typography>
+      <Typography
+        ff="PRETENDARD"
+        fs={32}
+        lh={48}
+        fw="SEMIBOLD"
+        color="BLACK"
+        style={{ width: "fit-content" }}
+      >
         동아리연합회 통합 플랫폼,{" "}
         <span style={{ color: colors.PRIMARY }}>Clubs</span>입니다
-      </PageTitle>
+      </Typography>
     </PageTitleWrapper>
 
     <NoticeAndServiceWrapper>
